fix: 서버 값 관련 데이터 타입 변경으로 인한 타입 변경

+++ O-G-Market/Models/Request/ProductUpdate.swift
@@ -11,9 +11,9 @@ struct ProductUpdate: Encodable {
     let name: String?
     let descriptions: String?
     let thumbnailID: Int?
-    let price: Int?
+    let price: Double?
     let currency: Currency?
-    let discountedPrice: Int?
+    let discountedPrice: Double?
     let stock: Int?
     let secret: String
